Hyper-V: fix tgt iSCSI targets disconnect issue
When detaching volumes, the Hyper-V driver decides whether the
iSCSI target should be disconnected or not by comparing the number
of volumes being disconnected and the total number of available LUNs
of the according iSCSI target.

This causes problems as tgt exports one extra reserved LUN. For
this reason, the according iSCSI target won't be disconnected
when a volume exported by TGT is detached.

The fix consists in checking the disk types when counting the
available LUNs of a target.

Change-Id: I0c2a723b94dae5b1c3838b9644537324606128e5
Closes-Bug: #1395063

Signed-off-by: Luis Fernandez Alvarez <luis.fernandez.alvarez@cern.ch>