Hyper-V: Fixes wrong hypervisor_version
The hypervisor_version will be stored in the database as
an Integer and it will be used by the scheduler, if required
by the image property 'hypervisor_version_requires'.

The hypervisor_version will then be converted back to a version
by spliting the int in groups of 3 digits.
E.g.: hypervisor_version 6003 is converted to '6.3'.

Currently, the reported hypervisor_version by the Hyper-V driver
is wrong. For example, for Windows Hyper-V / Server 2012 R2
or Windows 8.1, the hypervisor_version reported is '639600',
which will later be converted to 639.600 by the scheduler.

Closes-Bug: #1409258

Change-Id: Id991fefed08d1da2a68db4c27de72ac2252b1781

Signed-off-by: Luis Fernandez Alvarez <luis.fernandez.alvarez@cern.ch>