OS-4778: add os, os_version in port extra_dhcp_opts
In the case of an instance created from an image but booted from volume
we do not get the appropriate metadata when creating the port - we need
os and os_version to be passed to landb.

The reason is that port creation happens before the block device
mapping is done, so we need access to extra information in the db.

Before creating the port, fetch the os and os_version as pass them to
the port creation via extra_dhcp_opts.